pyproject.toml
[tool.poetry]
name = "ticket-management-system"
version = "0.1.0"
description = "A simple ticket management system project"
authors = ["Aayam Maharjan <aayamm.calcgen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.4"
djangorestframework = "^3.15.1"
django-environ = "^0.12.0"
django-debug-toolbar = "^6.0.0"
drf-yasg = "^1.21.10"
psycopg2-binary = "^2.9.10"
django-redis = "^6.0.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.1.0"}
djangorestframework-simplejwt = "^5.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
